pyproject.toml
[project]
name = "nrgpt-iclr"
version = "0.1.0"
description = "NRGPT submission to ICLR"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "attrs>=25.1.0",
    "cattrs>=25.2.0",
    "datasets<4.0",
    "einops>=0.8.1",
    "fastapi>=0.100.0",
    "fvcore>=0.1.5.post20221221",
    "importlib-metadata>=6.0.0",
    "jaxtyping>=0.2.38",
    "matplotlib>=3.7.0",
    "nbformat>=5.10.4",
    "numpy<2.0",
    "openpyxl>=3.1.5",
    "opt-einsum>=3.4.0",
    "polars>=1.33.1",
    "pyarrow>=12.0.0,<15.0.0",
    "pydantic>=2.0.0",
    "pyspellchecker>=0.8.3",
    "python-multipart>=0.0.6",
    "pyyaml>=6.0",
    "sentence-transformers>=5.1.0",
    "starlette>=0.27.0",
    "tabulate>=0.9.0",
    "textstat>=0.7.10",
    "tiktoken>=0.9.0",
    "torch>=2.6.0",
    "tqdm>=4.67.1",
    "transformers>=4.49.0",
    "tyro>=0.9.14",
    "typing-extensions>=4.5.0",
    "uvicorn>=0.22.0",
    "wandb>=0.19.6",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.0",
    "pytest>=8.3.5",
]
